requirements.txt
python==3.10.9
faiss-cpu
langchain
google-generativeai
PyPDF2>=2.0